+++ core-ng/src/test/java/core/framework/internal/log/ActionLogTest.java
@@ -122,23 +122,11 @@ class ActionLogTest {
 
     @Test
     void trace() {
-        String suffix = "...(soft trace limit reached)\n";
-        String trace = log.trace(200, 500);
-        assertThat(trace).hasSize(200 + suffix.length())
-            .contains("ActionLog - begin")
-            .endsWith(suffix);
+        String trace = log.trace();
+        assertThat(trace).contains("c.f.i.log.ActionLog - begin");
 
         log.process(new LogEvent("logger", null, LogLevel.WARN, "warning", null, null));
-
-        trace = log.trace(200, 500);
-        assertThat(trace).endsWith("warning\n");
-
-        trace = log.trace(250, 500);   // the max debug length will hit warning event
-        assertThat(trace).contains("warning")
-            .endsWith(suffix);
-
-        log.process(new LogEvent("logger", null, LogLevel.WARN, "warning2", null, null));
-        trace = log.trace(250, 320);   // truncate with hard limit
-        assertThat(trace).endsWith("...(hard trace limit reached)");
+        trace = log.trace();
+        assertThat(trace).contains("WARN logger - warning");
     }
 }